Close to working. I still need to check the grid interpolation is being done correctly.
